pyproject.toml
[tool.poetry]
name = "mastodon-api"
version = "0.1.0"
description = "A API to studies."
authors = ["Pery Lemke <pery.lemke@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.16.0"
black = "^21.12b0"
autopep8 = "^1.6.0"
python-decouple = "^3.5"
dnspython = "^2.1.0"
pymongo = "3.12"
motor = "^2.5.1"
requests = "^2.26.0"
pytest-asyncio = "^0.16.0"
httpx = "^0.21.1"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
